backend/requirements.txt
# Backend Mental Coaching API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.28.0
pinecone-client>=3.2.0

